Make Forge lanes cover all dapps

+++ dapps/bike-nft/src/BicycleComponentOpsFund.sol
@@ -23,14 +23,14 @@ contract BicycleComponentOpsFund is AccessControl {
     bytes32 public constant OPS_MANAGER_ROLE = keccak256("OPS_MANAGER_ROLE");
     bytes32 public constant CARTE_BLANCHE_ROLE = keccak256("CARTE_BLANCHE_ROLE");
 
-    uint public defaultAllowanceIncrement = 3;
+    uint256 public defaultAllowanceIncrement = 3;
 
-    mapping(address => uint) private _allowanceOf;
+    mapping(address => uint256) private _allowanceOf;
 
-    event DefaultAllowanceIncrementSet(uint newInitialOpsAllowance);
-    event AllowanceAdded(address indexed owner, uint amount);
-    event AllowanceConsumed(address indexed owner, uint amount);
-    event AllowanceRefunded(address indexed owner, uint amount);
+    event DefaultAllowanceIncrementSet(uint256 newInitialOpsAllowance);
+    event AllowanceAdded(address indexed owner, uint256 amount);
+    event AllowanceConsumed(address indexed owner, uint256 amount);
+    event AllowanceRefunded(address indexed owner, uint256 amount);
 
     constructor() {
         _setupRole(DEFAULT_ADMIN_ROLE, msg.sender);
@@ -39,16 +39,16 @@ contract BicycleComponentOpsFund is AccessControl {
         _setupRole(CARTE_BLANCHE_ROLE, msg.sender);
     }
 
-    function allowanceOf(address owner) public view returns (uint) {
-        return hasRole(CARTE_BLANCHE_ROLE, owner) ? type(uint).max : _allowanceOf[owner];
+    function allowanceOf(address owner) public view returns (uint256) {
+        return hasRole(CARTE_BLANCHE_ROLE, owner) ? type(uint256).max : _allowanceOf[owner];
     }
 
-    function setDefaultAllowanceIncrement(uint newDefaultAllowanceIncrement) public onlyRole(DEFAULT_ADMIN_ROLE) {
+    function setDefaultAllowanceIncrement(uint256 newDefaultAllowanceIncrement) public onlyRole(DEFAULT_ADMIN_ROLE) {
         defaultAllowanceIncrement = newDefaultAllowanceIncrement;
         emit DefaultAllowanceIncrementSet(newDefaultAllowanceIncrement);
     }
 
-    function addAllowance(address owner, uint amount) public onlyRole(OPS_MANAGER_ROLE) {
+    function addAllowance(address owner, uint256 amount) public onlyRole(OPS_MANAGER_ROLE) {
         if (!hasRole(CARTE_BLANCHE_ROLE, owner)) {
             _allowanceOf[owner] += amount;
         }
@@ -56,7 +56,7 @@ contract BicycleComponentOpsFund is AccessControl {
         emit AllowanceAdded(owner, amount);
     }
 
-    function consume(address owner, uint amount) public onlyRole(PAYMASTER_ROLE) {
+    function consume(address owner, uint256 amount) public onlyRole(PAYMASTER_ROLE) {
         if (!hasRole(CARTE_BLANCHE_ROLE, owner)) {
             require(_allowanceOf[owner] >= amount, "Insufficient ops allowance");
             _allowanceOf[owner] -= amount;
@@ -65,7 +65,7 @@ contract BicycleComponentOpsFund is AccessControl {
         emit AllowanceConsumed(owner, amount);
     }
 
-    function refund(address owner, uint amount) public onlyRole(PAYMASTER_ROLE) {
+    function refund(address owner, uint256 amount) public onlyRole(PAYMASTER_ROLE) {
         if (!hasRole(CARTE_BLANCHE_ROLE, owner)) {
             _allowanceOf[owner] += amount;
         }
