+++ dapps/bike-nft/src/Utils.sol
@@ -2,28 +2,40 @@ pragma solidity 0.8.35;
 
 import "@openzeppelin/contracts-upgradeable/utils/Base64Upgradeable.sol";
 
-
 library Utils {
-    function stringifyOnChainMetadata(string memory spacer, string memory name, string memory description, string memory imageURL, string[] memory traitTypes, string[] memory traitValues)
-    internal pure returns (string memory)
-    {
+    function stringifyOnChainMetadata(
+        string memory spacer,
+        string memory name,
+        string memory description,
+        string memory imageURL,
+        string[] memory traitTypes,
+        string[] memory traitValues
+    ) internal pure returns (string memory) {
         string memory jsonBody = string(
             abi.encodePacked(
-                '{',
-                '"name":"', escapeJSONString(name), '",', spacer,
-                '"description":"', escapeJSONString(description), '",', spacer,
-                '"image":"', escapeJSONString(imageURL), '"'
+                "{",
+                '"name":"',
+                escapeJSONString(name),
+                '",',
+                spacer,
+                '"description":"',
+                escapeJSONString(description),
+                '",',
+                spacer,
+                '"image":"',
+                escapeJSONString(imageURL),
+                '"'
             )
         );
 
         require(traitTypes.length == traitValues.length, "Type/Value arrays must have the same length");
 
         if (traitTypes.length > 0) {
-            jsonBody = string(abi.encodePacked(jsonBody, ',', spacer, '"attributes":['));
+            jsonBody = string(abi.encodePacked(jsonBody, ",", spacer, '"attributes":['));
 
-            for (uint i = 0; i < traitTypes.length; i++) {
+            for (uint256 i = 0; i < traitTypes.length; i++) {
                 if (i > 0) {
-                    jsonBody = string(abi.encodePacked(jsonBody, ','));
+                    jsonBody = string(abi.encodePacked(jsonBody, ","));
                 }
 
                 jsonBody = string(
@@ -40,21 +52,19 @@ library Utils {
                 );
             }
 
-            jsonBody = string(abi.encodePacked(jsonBody, ']'));
+            jsonBody = string(abi.encodePacked(jsonBody, "]"));
         }
 
-        jsonBody = string(abi.encodePacked(jsonBody, '}'));
+        jsonBody = string(abi.encodePacked(jsonBody, "}"));
 
         return jsonBody;
     }
 
-    function escapeJSONString(string memory value)
-    internal pure returns (string memory)
-    {
+    function escapeJSONString(string memory value) internal pure returns (string memory) {
         bytes memory input = bytes(value);
         bytes memory output;
 
-        for (uint i = 0; i < input.length; i++) {
+        for (uint256 i = 0; i < input.length; i++) {
             bytes1 char = input[i];
 
             if (char == bytes1(0x22)) {
@@ -81,20 +91,11 @@ library Utils {
         return string(output);
     }
 
-    function _hexDigit(uint8 value)
-    private pure returns (bytes1)
-    {
+    function _hexDigit(uint8 value) private pure returns (bytes1) {
         return value < 10 ? bytes1(value + 0x30) : bytes1(value + 0x57);
     }
 
-    function packJSON(string memory jsonString)
-    internal pure returns (string memory)
-    {
-        return string(
-            abi.encodePacked(
-                "data:application/json;base64,",
-                Base64Upgradeable.encode(bytes(jsonString))
-            )
-        );
+    function packJSON(string memory jsonString) internal pure returns (string memory) {
+        return string(abi.encodePacked("data:application/json;base64,", Base64Upgradeable.encode(bytes(jsonString))));
     }
 }
